requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
jinja2==3.1.4
pandas==2.2.2
openpyxl==3.1.5
numpy==1.26.4
python-dotenv==1.0.1
groq==0.9.0
gunicorn==22.0.0